+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 1.0.5
 - NUCLEAR and ghost options
+- You can change the color of the pips in settings.